Lab_Sesion3/requirements.txt
# Dependencias - Sesion 3: Bases de Datos
# Instalar: pip install -r requirements.txt

# Framework web
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3

# Base de datos
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Utilidades
python-dotenv>=1.0.0

# Testing
requests>=2.31.0
pytest>=8.0.0
